Cleaned a bit the code to remove LV reference, and have all pods behavior on one example page
