+++ Example/Stickers/Base.lproj/Main.storyboard
@@ -41,13 +41,6 @@
                                     <constraint firstItem="02w-sb-0w2" firstAttribute="centerX" secondItem="Mzo-7l-C2j" secondAttribute="centerX" id="znD-0B-1CY"/>
                                 </constraints>
                             </view>
-                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="SV4-st-Pd2">
-                                <rect key="frame" x="150" y="30" width="75" height="30"/>
-                                <state key="normal" title="Button"/>
-                                <connections>
-                                    <action selector="changePage:" destination="vXZ-lx-hvc" eventType="touchUpInside" id="8vO-bm-IsF"/>
-                                </connections>
-                            </button>
                             <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="cp6-D2-7cj">
                                 <rect key="frame" x="16" y="20" width="85" height="30"/>
                                 <state key="normal" title="Add stickers"/>
@@ -64,10 +57,7 @@
                             </button>
                         </subviews>
                         <constraints>
-                            <constraint firstItem="SV4-st-Pd2" firstAttribute="centerX" secondItem="WDB-Zo-fkc" secondAttribute="centerX" id="3TT-oc-PJt"/>
-                            <constraint firstItem="SV4-st-Pd2" firstAttribute="leading" secondItem="WDB-Zo-fkc" secondAttribute="leading" constant="150" id="3aR-gh-nOx"/>
                             <constraint firstItem="WDB-Zo-fkc" firstAttribute="bottom" secondItem="Mzo-7l-C2j" secondAttribute="bottom" id="4XG-4V-N17"/>
-                            <constraint firstItem="SV4-st-Pd2" firstAttribute="top" secondItem="WDB-Zo-fkc" secondAttribute="top" constant="10" id="4aX-3e-x97"/>
                             <constraint firstItem="WDB-Zo-fkc" firstAttribute="trailing" secondItem="Mzo-7l-C2j" secondAttribute="trailing" id="5dw-sh-Xcg"/>
                             <constraint firstItem="WDB-Zo-fkc" firstAttribute="trailing" secondItem="lvn-p7-iuk" secondAttribute="trailing" constant="10" id="Cty-aL-OUL"/>
                             <constraint firstItem="lvn-p7-iuk" firstAttribute="top" secondItem="WDB-Zo-fkc" secondAttribute="top" constant="10" id="FN6-FJ-g7D"/>
@@ -86,38 +76,6 @@
                 <placeholder placeholderIdentifier="IBFirstResponder" id="x5A-6p-PRh" sceneMemberID="firstResponder"/>
             </objects>
             <point key="canvasLocation" x="141.59999999999999" y="133.5832083958021"/>
-        </scene>
-        <!--Gif Creator-->
-        <scene sceneID="A1L-5f-msg">
-            <objects>
-                <viewController storyboardIdentifier="GifCreator" id="sF9-3u-r2H" customClass="GifCreator" customModule="Stickers_Example" customModuleProvider="target" sceneMemberID="viewController">
-                    <layoutGuides>
-                        <viewControllerLayoutGuide type="top" id="cVt-73-VV1"/>
-                        <viewControllerLayoutGuide type="bottom" id="CB0-z6-Fja"/>
-                    </layoutGuides>
-                    <view key="view" contentMode="scaleToFill" id="RQY-8M-eTk">
-                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-                        <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-                        <subviews>
-                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="YgO-j0-PDP">
-                                <rect key="frame" x="154" y="52" width="67" height="30"/>
-                                <state key="normal" title="Button"/>
-                                <connections>
-                                    <action selector="takeScreen:" destination="sF9-3u-r2H" eventType="touchUpInside" id="md8-HO-ra2"/>
-                                </connections>
-                            </button>
-                        </subviews>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                        <constraints>
-                            <constraint firstItem="YgO-j0-PDP" firstAttribute="top" secondItem="cVt-73-VV1" secondAttribute="bottom" constant="32" id="Luy-P0-Fii"/>
-                            <constraint firstItem="YgO-j0-PDP" firstAttribute="leading" secondItem="RQY-8M-eTk" secondAttribute="leading" constant="154" id="Yix-2c-EvJ"/>
-                            <constraint firstItem="YgO-j0-PDP" firstAttribute="centerX" secondItem="RQY-8M-eTk" secondAttribute="centerX" id="ziU-er-LN2"/>
-                        </constraints>
-                    </view>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="IsO-nP-vQ0" userLabel="First Responder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="820" y="134"/>
         </scene>
     </scenes>
 </document>
